app/pyproject.toml
[project]
name = "navigator_backend"
version = "1.26.1"
description = ""
authors = [{ name = "CPR-dev-team", email = "tech@climatepolicyradar.org" }]
requires-python = ">=3.11"
dependencies = [
  "Authlib>=1.6.0",
  "bcrypt>=4.3.0",
  "boto3>=1.26",
  "cpr-sdk[vespa,datasets]==1.20.1",
  "fastapi>=0.115.8",
  "fastapi-health>=0.4.0",
  "fastapi-pagination[sqlalchemy]>=0.13.2",
  "httpx>=0.28.1",
  "itsdangerous>=2.1.0",
  "json-logging>=1.3.0",
  "pandas>=2.3.0",
  "passlib>=1.7.4",
  "psycopg2>=2.9.10",
  "PyJWT>=2.3.0",
  "python-multipart>=0.0.20",
  "python-slugify>=6.1.2",
  "requests>=2.32",
  "requests-toolbelt>=1.0.0",
  "SQLAlchemy>=1.4.31",
  "SQLAlchemy-Utils>=0.38.2",
  "starlette==0.40.0",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.34.3",
  "botocore>=1.38.37",
  "db-client @ git+https://github.com/climatepolicyradar/navigator-db-client.git@v3.9.17",
  "urllib3<3",
  "apscheduler>=3.10.4",
  "numpy==1.26.4",
  "python-dateutil>=2.9.0.post0",
  "opentelemetry-instrumentation-fastapi>=0.54b0",
  "opentelemetry-sdk>=1.33.0",
  "python-dotenv>=1.1.0",
  "pydantic-settings>=2.9.1",
  "opentelemetry-exporter-otlp>=1.33.0",
  "opentelemetry-instrumentation-sqlalchemy>=0.54b0",
]

[project.optional-dependencies]
dev = [
  "black>=25.1.0",
  "moto[s3]>=5.1.6",
  "pytest>=8.4.0",
  "pytest-asyncio>=1.0.0",
  "pytest-mock>=3.14.1",
  "pyright>=1.1.402",
  "ruff>=0.11.13",
  "types-SQLAlchemy>=1.4.31",
  "surrogate>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
testpaths = ["tests"]

markers = ["cors", "search", "unit"]
asyncio_mode = "strict"

[tool.pydocstyle]
ignore = """
"""

[tool.pyright]
include = ["./"]
exclude = [
  "**/__pycache__",
  "concepts-api/**",
  "families-api/**",
  "geographies-api/**",
  "api/**",
]
ignore = ["scripts/**/*"]
pythonVersion = "3.10"
pythonPlatform = "Linux"
venvPath = ".venv"

[virtualenvs]
in-project = true
path = ".venv"
